code cleanup, open vod folder button
