+++ src-tauri/capabilities/default.json
